@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "private": true,
   "dependencies": {
+    "@reduxjs/toolkit": "^1.8.4",
     "autosize": "^5.0.0",
     "axios": "^0.21.1",
     "bootstrap": "^5.0.1",
@@ -12,15 +13,12 @@
     "jquery": "^3.6.0",
     "jsonwebtoken": "^8.5.1",
     "marked": "^4.0.12",
-    "react": "^17.0.2",
-    "react-device-detect": "^1.17.0",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
     "react-transition-group": "^4.4.2",
-    "redux": "^4.1.0",
-    "redux-thunk": "^2.3.0",
     "sass": "^1.34.1",
     "web-vitals": "^1.1.2"
   },
@@ -51,7 +49,6 @@
   },
   "devDependencies": {
     "nodemon": "^2.0.12",
-    "redux-devtools-extension": "^2.13.9",
     "terser": "^3.14.1"
   }
 }
