@@ -1,17 +1,17 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import {Router} from "react-router-dom";
+import { createRoot } from 'react-dom/client';
 import App from "./App";
 
-import {Provider} from "react-redux";
+import { Provider } from "react-redux";
 import store from "./store";
-import history from "./helpers/history"
 
-ReactDOM.render(
-    <Provider store={store}>
-        <Router history={history}>
-            <App />
-        </Router>
-    </Provider>,
-    document.getElementById('root')
+const container = document.getElementById('root');
+const root = createRoot(container); 
+
+root.render(
+	<React.StrictMode>
+		<Provider store={store}>
+			<App />
+		</Provider>
+	</React.StrictMode>
 );
