@@ -1,5 +1,10 @@
-import React, { useEffect } from "react";
-import { Route, Switch } from "react-router-dom";
+import React, { useEffect, useState } from "react";
+import {
+	BrowserRouter as Router,
+	Route,
+	Routes
+  } from "react-router-dom";
+// import { Route, Switch } from "react-router-dom";
 import history from "./helpers/history";
 // import jwt from 'jsonwebtoken';
 import $ from 'jquery';
@@ -12,8 +17,8 @@ import NavBar from "./Components/navBar";
 
 
 import Home from './pages/Home'
-import About from './pages/About'
-import Works from './pages/Work'
+import About from './pages/About';
+import Works from './pages/Work';
 import Error from './pages/ErrorPage'
 import Contact from "./pages/Contact";
 import Admin from "./pages/Admin";
@@ -29,6 +34,10 @@ import {
 } from "react-transition-group";
 
 import { useSelector } from "react-redux";
+import Project from "./pages/Project";
+import SharedLayout from "./Components/SharedLayout";
+import ProtectetdRoute from "./Components/ProtectedRoute";
+import SharedProductLayout from "./Components/ShareProductLayout";
 // const routes = [
 //     { path: '/', Component: Home },
 //     { path: '/about', Component: About },
@@ -62,34 +71,26 @@ import { useSelector } from "react-redux";
 // );
 function App() {
     // const location = useLocation();
-    const { isOpen } = useSelector(state => state.hamburger);
+    // const { isOpen } = useSelector(state => state.hamburger);
+	const [user, setUser] = useState(null);
 
-    const disableScroll = () => {
-        $("body").css("overflow", "hidden");
-        $('.page').css('opacity', '0.1');
-
-    }
-    const enableScroll = () => {
-        $("body").css("overflow", "auto");
-        $('.page').css('opacity', '1');
-    }
     useEffect(() => {
-        console.log("App gestarted ...")
+		console.log("App gestarted ...")
         // const mainStorage = getFromStorage("session");
         // add the session
         // if (!mainStorage) {
-        //     SERVER_Request.get('/api/session').then((response) => {
-        //         if (response.data.session) {
+			//     SERVER_Request.get('/api/session').then((response) => {
+				//         if (response.data.session) {
         //             setInStorage("session", response.data.session);
         //         }
         //     }).then((response) => {
         //         console.log("Add sessio response: \n", response);
         //     })
         // } else {
-        //     // delete the session 
-        //     const exp = jwt.decode(mainStorage);
-        //     if (Date.now() >= (exp.exp * 1000)) {
-        //         removeInStorage("session");
+			//     // delete the session 
+			//     const exp = jwt.decode(mainStorage);
+			//     if (Date.now() >= (exp.exp * 1000)) {
+				//         removeInStorage("session");
         //         SERVER_Request.post(`/api/session/${exp.id}`)
         //             .then((response) => {
         //                 console.log(response.message);
@@ -97,68 +98,56 @@ function App() {
         //     }
         // }
     }, [])
+	
+	// const disableScroll = () => {
+	// 	$("body").css("overflow", "hidden");
+	// 	$('.page').css('opacity', '0.1');
+
+	// }
+	// const enableScroll = () => {
+	// 	$("body").css("overflow", "auto");
+	// 	$('.page').css('opacity', '1');
+	// }
 
     // if (isOpen) {
-    //     disableScroll();
-    // } else {
-    //     enableScroll();
-    // }
+		//     disableScroll();
+		// } else {
+			//     enableScroll();
+			// }
     return (
         <div id="app">
-            <NavBar />
-            <SideBar />
-
-            <TransitionGroup>
+			<Router>
+				<Routes>
+					<Route path="/" element={ <SharedLayout/> }>
+						<Route index element={ <Home />} />
+						<Route path="works" element={ <SharedProductLayout />}> 
+							<Route index element={<Works />} />
+							<Route path=":projectName" element={ <Project />} />
+						</Route>
+						<Route path="/contact" element={ <Contact />} />
+						<Route path="/about" element={ <About /> } />
+						<Route 
+							path="/admin"
+							element = {
+								<ProtectetdRoute user={user} >
+									<Admin user={user} />
+							 	</ProtectetdRoute>
+							} 
+						/>
+						<Route path="/login" element={ <LogInOut setUser={setUser} /> } />
+						<Route path="*" element={ <Error /> } />
+					</Route>
+				</Routes>
+			</Router>
+            {/* <TransitionGroup>
                 <CSSTransition
                     timeout={1000}
                     classNames="pages"
-                    key={history.location.key}>
-                    <Switch location={history.location}>
-                        <Route exact path="/sign">
-                            <LogInOut history={history} />
-                        </Route>
-                        <Route exact path="/">
-                            <Home />
-                        </Route>
-
-                        <Route path="/about">
-                            <About />
-                        </Route>
-
-                        <Route path="/contact">
-                            <Contact />
-                        </Route>
-
-                        <Route path="/work">
-                            <Works />
-                        </Route>
-
-                        <Route path="/admin">
-                            <Admin />
-                        </Route>
-
-                        <Route path="*">
-                            <Error />
-                        </Route>
-                    </Switch>
-                </CSSTransition>
-            </TransitionGroup>
-            {/* </BrowserView>
-            <MobileView>
-                <Switch>
-                    <Route exact path="/">
-                        <NavBar />
-                        <Home />
-                        <About />
-                        <Works />
-                        <Contact />
-                    </Route>
-                    <Route exact path="/sign">
-                        <NavBar />
-                        <Sign />
-                    </Route>
-                </Switch>
-            </MobileView> */}
+                    key={history.location.key}> */}
+                    {/* <Switch location={history.location}>
+                    </Switch> */}
+                {/* </CSSTransition>
+            </TransitionGroup> */}
         </div>
     );
 }
