@@ -1,11 +1,12 @@
 import React from "react";
-import { NavLink } from "react-router-dom";
+import { Link, NavLink } from "react-router-dom";
 // import { BrowserView, MobileView } from 'react-device-detect';
-import { toggleHamburger } from "../redux/actions/hamburger";
 import { useSelector, useDispatch } from "react-redux";
 // import history from "../helpers/history";
 import { Logo } from "./logo";
 import Hamburger from "./hamburger";
+import { toggleHamburger } from '../features/hamburger/hamburgerSlice';
+
 
 export const NavBar = (props) => {
     const { isOpen } = useSelector(state => state.hamburger);
@@ -21,9 +22,11 @@ export const NavBar = (props) => {
             <nav className="nav">
                 <ul>
                     <li className="nav-logo">
-                        <NavLink activclassname="active"
+                        <NavLink 
+						    // activclassname="active"
                             to='/'
-                            onClick={handleLogoClick}>
+                            onClick={handleLogoClick}
+						>
                             <Logo />
                         </NavLink>
                     </li>
