@@ -1,27 +1,15 @@
 import React from 'react';
-// import { BrowserView, MobileView } from 'react-device-detect';
-
-import { NavLink } from 'react-router-dom'
+import { Link, NavLink } from 'react-router-dom';
 
 // From Redux
 import { useSelector, useDispatch } from "react-redux";
-import { toggleHamburger } from "../redux/actions/hamburger"
 import { Switch } from "./switch"
-
-// Form React Scroll
-
-// From React CSS Transition
+import { toggleHamburger } from '../features/hamburger/hamburgerSlice';
 import CSSTransition from "react-transition-group/cjs/CSSTransition";
 
-
-// const TOP_BOTTOM = "showDown";
-// const TOP_TOP = "showUp";
 const SideBar = () => {
-    // Get the actual Hamburger state from Redux Store
     const { isOpen } = useSelector(state => state.hamburger);
     const dispatch = useDispatch();
-
-
     // const [curVanishingMode, setCurVanishingMode] = useState("showUp")
 
     const handleSwitchClick = () => {
@@ -51,7 +39,6 @@ const SideBar = () => {
     // }
     return (
         <>
-            {/* <BrowserView> */}
             <CSSTransition
                 in={isOpen}
                 classNames="show"
@@ -63,9 +50,12 @@ const SideBar = () => {
                             <NavLink
                                 title="home"
                                 className="side__link"
-                                activclassname="active"
                                 exact to='/'
-                                onClick={handleSwitchClick}>
+                                onClick={handleSwitchClick}
+								// style={({isActive}) => {
+								// 	return { backgroundColor: isActive ? "red" : "green"}
+								// }}
+								>
                                 <i className="fa fa-home" />
                             </NavLink>
                         </li>
@@ -73,7 +63,6 @@ const SideBar = () => {
                             <NavLink
                                 title="about"
                                 className="side__link"
-                                activclassname="active"
                                 to='/about'
                                 onClick={handleSwitchClick}>
                                 <i className="fa fa-id-card " />
@@ -83,8 +72,7 @@ const SideBar = () => {
                             <NavLink
                                 title="work"
                                 className="side__link"
-                                activclassname="active"
-                                to='/work'
+                                to='/works'
                                 onClick={handleSwitchClick}>
                                 <i className="fa fa-briefcase " />
                             </NavLink>
@@ -93,7 +81,6 @@ const SideBar = () => {
                             <NavLink
                                 title="contact"
                                 className="side__link"
-                                activclassname="active"
                                 to='/contact'
                                 onClick={handleSwitchClick}>
                                 <i className="fa fa-address-card" />
